lenovo: add hybrid graphics driver
Add a universal hybrid graphics driver compatible with
all supported lenovo devices.
Hybrid graphics allows to internaly connect the display panel to
either of one GPUs.
As there are only two GPUs one GPIO needs to be toggled.
In case the discrete GPU is activated the panel is routed to it.
On deactivation the panel is routed to the integrated
GPU and the discrete HDMI audio device is disabled too.

Change-Id: I9b80b31a7749bdf893ed3b772a6505c9f29a56d1